+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Shorten message content preview for highlight DMs, embed remains unaffected
 - Starboard embeds now show the timestamp of the starred message's creation in the footer
 - Starboard embeds now show a preview of replied-to messages if applicable
+- Starboard now shows the server-specific display name and avatar of the user starred (and replied-to user if relevant)
 
 ## Removals